pancake sorting algorithm
Unlike a traditional sorting algorithm, which attempts to sort with the fewest comparisons possible, the goal is to sort the sequence in as few reversals as possible. 
The idea is to do something similar to Selection Sort. We one by one place maximum element at the end and reduce the size of current array by one. This algorithm implements the pancake sorting algorithm